+++ classes/class-base.php
@@ -196,9 +196,16 @@ class Base {
 	 * @return array
 	 */
 	public function add_action_links( $actions ) {
-		$action_link = [ '<a href="' . admin_url( 'admin.php?page=progress-planner' ) . '">' . __( 'Dashboard', 'progress-planner' ), '</a>' ];
-		$actions     = array_merge( $action_link, $actions );
-		return $actions;
+		return array_merge(
+			[
+				sprintf(
+					'<a href="%1$s">%2$s</a>',
+					admin_url( 'admin.php?page=progress-planner' ),
+					__( 'Dashboard', 'progress-planner' )
+				),
+			],
+			$actions
+		);
 	}
 
 	/**