+++ classes/suggested-tasks/local-tasks/providers/class-disable-comments.php
@@ -1,0 +1,77 @@
+<?php
+/**
+ * Add tasks for settings saved.
+ *
+ * @package Progress_Planner
+ */
+
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+
+/**
+ * Add tasks to check if WP debug is enabled.
+ */
+class Disable_Comments extends Local_OneTime_Tasks_Abstract {
+
+	/**
+	 * The provider type.
+	 *
+	 * @var string
+	 */
+	const TYPE = 'configuration';
+
+	/**
+	 * The provider ID.
+	 *
+	 * @var string
+	 */
+	const ID = 'disable-comments';
+
+	/**
+	 * Check if the task condition is satisfied.
+	 * (bool) true means that the task condition is satisfied, meaning that we don't need to add the task or task was completed.
+	 *
+	 * @return bool
+	 */
+	public function should_add_task() {
+		return 10 < \wp_count_comments()->approved && 'closed' !== \get_option( 'disable_comments' );
+	}
+
+	/**
+	 * Check if the task is completed.
+	 *
+	 * @return bool
+	 */
+	public function is_task_completed() {
+		return 'closed' === \get_option( 'disable_comments' );
+	}
+
+	/**
+	 * Get the task details.
+	 *
+	 * @param string $task_id The task ID.
+	 *
+	 * @return array
+	 */
+	public function get_task_details( $task_id = '' ) {
+
+		if ( ! $task_id ) {
+			$task_id = $this->get_provider_id();
+		}
+
+		return [
+			'task_id'     => $task_id,
+			'title'       => \esc_html__( 'Disable comments', 'progress-planner' ),
+			'parent'      => 0,
+			'priority'    => 'high',
+			'type'        => $this->get_provider_type(),
+			'points'      => 1,
+			'url'         => $this->capability_required() ? \esc_url( \admin_url( 'options-discussion.php' ) ) : '', // @phpstan-ignore-line property.nonObject
+			'dismissable' => true,
+			'description' => '<p>' . sprintf(
+				// translators: %d is the number of approved comments.
+				\esc_html__( 'There are %d comments. If you don\'t need comments on your site, consider disabling them.', 'progress-planner' ),
+				\wp_count_comments()->approved,
+			) . '</p>',
+		];
+	}
+}