+++ classes/suggested-tasks/local-tasks/providers/class-hello-world.php
@@ -41,14 +41,13 @@ class Hello_World extends Local_OneTime_Tasks_Abstract {
 	protected $sample_post = false;
 
 	/**
-	 * Check if the task condition is met.
+	 * Check if the task condition is satisfied.
+	 * (bool) true means that the task condition is satisfied, meaning that we don't need to add the task or task was completed.
 	 *
 	 * @return bool
 	 */
-	public function check_task_condition() {
-		$hello_world = $this->get_sample_post();
-
-		return null === $hello_world ? true : false;
+	public function should_add_task() {
+		return null !== $this->get_sample_post();
 	}
 
 	/**