+++ classes/suggested-tasks/local-tasks/providers/class-settings-saved.php
@@ -27,14 +27,12 @@ class Settings_Saved extends Local_OneTime_Tasks_Abstract {
 	const ID = 'settings-saved';
 
 	/**
-	 * Check if the task condition is met.
+	 * Check if the task should be added.
 	 *
 	 * @return bool
 	 */
-	public function check_task_condition() {
-		$prpl_pro_license_key = \get_option( 'progress_planner_pro_license_key', false );
-
-		return false !== $prpl_pro_license_key ? true : false;
+	public function should_add_task() {
+		return false === \get_option( 'progress_planner_pro_license_key', false );
 	}
 
 	/**