+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+= 1.0.5 =
+
+Under the hood:
+
+* Improved suggested tasks completion conditions.
+* Improved checks for suggested 'review post' tasks.
+
+We've added the following Recommendations from Ravi:
+
+* [Update PHP version](TBD).
+* [Disable comments on your site](TBD).
+
 = 1.0.4 =
 
 Enhancements: